refactors and cleanup
